wip: new implementation of atom
