+++ deprecated/index.d.ts
@@ -1,4 +1,9 @@
-import { WritableStore, ReadableStore, StoreValue } from '../atom/index.js'
+import type {
+  WritableStore,
+  ReadableStore,
+  StoreValue
+} from '../atom/depricated.js'
+
 import { MapBuilder } from '../map-template/index.js'
 import { MapStore } from '../map/index.js'
